Clean up the styling for the resources pages Communities, Icons, and Websites.
The styling for these was a bit inconsistent and generally not
particularly pretty, so I made the image for each link aligned to
the far right of its section. Also added a bit more margins to the
descriptions of them on the left so that the border-left isn't right
next to the edge.